integrate tokenlist api for fuse
