+++ server/config/test.json
@@ -15,8 +15,5 @@
   "mongo": {
     "uri": "mongodb://localhost/CLN-community-app",
     "options": {}
-  },
-  "amazon": {
-    "apiBase": "https://s3-eu-west-1.amazonaws.com/cln-dapp-images-qa/"
   }
 }
