+++ server/src/routes/api/index.js
@@ -4,7 +4,7 @@ router.use('/metadata', require('./metadata'))
 router.use('/images', require('./images'))
 router.use('/receipts', require('./receipts'))
 router.use('/stats', require('./stats'))
-router.use('/balances', require('./balances'))
+router.use('/tokenlist', require('./tokenList'))
 router.use('/users', require('./users'))
 router.use('/tokens', require('./tokens'))
 router.use('/tokens/progress', require('./tokenProgress'))
